move entry clause resolution into model
